[unity.mvx] UILayer初始化优化，使用sortingLayerName来初始化

+++ mvvmcross_for_unity3d/Assets/Playground/Views/SplashScreenExWindow.cs
@@ -5,7 +5,7 @@ using UnityEngine.UI;
 
 namespace Playground.Views
 {
-    [MvxWindowPresentation((int)LAYER.top)]
+    [MvxWindowPresentation(LAYER.top)]
     public class SplashScreenExWindow : MvxUnityWindow<SplashScreeExViewModel>
     {
         public Button button;
